Moves parsers to conftest as session-scoped fixtures.
